rename(experimentation): _results_query -> results_query
